example: Apply workaround for GpioBinarySensorThing
GPIO undefined value on ready

gpio.js is bugged: value can be trusted before first change,
problem is know and fix is pending:

   Cant read initial value of pin #35

So until the upstream fix is released,
this workaround resolves the possibly wrong value on ready
(before 1st change).

This single change can be reverted on upgrade,
since previous base is sane.

It has been tested on RPi and ARTIK10.

Relate-to: https://github.com/EnotionZ/GpiO/issues/35
Change-Id: I2ea607f48e450a0728624ef15c17d14c98049072
Signed-off-by: Philippe Coval <p.coval@samsung.com>